requirements.txt
streamlit==1.25.0
pandas==1.5.3
Pillow==9.4.0